Implement search logic to retrieve relevant Arpa records for a given IP query.
A query can consist of single IPv4 or IPv6 addresses, or ranges in CIDR
notation. If an IPv4 range is given, as well as a single IP address that
falls within that range, only the range is retrieved. Retrieved IPv4
results are filtered so only IPs matching the queried ranges are
returned. This filtering does not take place for IPv6 ranges.