Fix IPv4 ARPA queries.
A query with a single (or multiple single) IPv4 addresses would result
in the entire relevant ARPA zone to be returned.

A query with a subnet would return only records of ARPA records it could
find. Now, all records in a given subnet are returned, with the proper
markings to show that the record is missing.