Add support for single transaction deletions and additions.
This has as added bonus that the RecordValidator can keep track of
records being deleted, so it can handle the situation that a user
deletes a CNAME, and adds an A record. Because the CNAME is deleted, the
addition of the A record is valid, because if the transaction is
committed, the zone will contain valid data.