Do not filter out NULL records when zone type is SLAVE
When we add an empty SLAVE zone, no records get inserted. The earlier
fix for PowerDNS 3.5 compatability introduced a bug where a freshly
created SLAVE zone can no longer be found or deleted, because no
records exist yet.

This fix revert to the old behaviour, but only in the cases of
SLAVE zones.